+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.1.0
+
+- Add support to show when packages are deprecated
+- Add support for versioning with "latest"
+
 ## 1.0.0
 
 Initial release